Attempt to get rid of circular dependency and other flaws. Results in weird AttributeError.
